Docstring + doctypes fix
