+++ rssscrapers/UK_dailymail_scraper.py
@@ -31,12 +31,16 @@ class dailymail(rss):
 
     def parsehtml(self,htmlsource):
         '''
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                                                                                                      
+        Parameters                                                                                                                                                               
+        ----                                                                                                                                                                       
+        htmlsource: string                                                                                                                                                                                                                                                                                                         
+        yields
+        ----
+        title    the title of the article
+        teaser    the intro to the artcile
         byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+        text    the plain text of the article
         '''
 
         try:
