+++ rssscrapers/UK_dailystar_scraper.py
@@ -32,11 +32,18 @@ class dailystar(rss):
     def parsehtml(self,htmlsource):
         '''
         Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
+
+
+        Parameters
+        ---- 
+        htmlsource: string
+
+        yields
+        ----                                                                                                        
+        title    the title of the article
+        teaser    the intro to the artcile
         byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+        text    the plain text of the article
         '''
 
         try:
