+++ rssscrapers/UK_theguardian_scraper.py
@@ -29,14 +29,23 @@ class theguardian(rss):
         self.version = ".1"
         self.date    = datetime.datetime(year=2017, month=9, day=13)
 
-    def parsehtml(self,htmlsource):
-        '''
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+    def parsehtml(self,htmlsource):    
+        '''                                                                                                                                                                                                                                                                 
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                 
+        
+
+        Parameters                                                                                                                                                                                                                                                         
+        ----                                                                                                                                                                                                                                                                
+        htmlsource: string                                                                                                                                                                                                                                                
+            html retrived from RSS feed                                                                                                                                                                                                                                     
+
+        yields                                                                                                                                                                                                                                                          
+        ----                                                                                                                                                                                                                                                                
+        title    the title of the article                                                                          
+        teaser    the intro to the artcile                                                                                                                                                             
+        byline    the author, e.g. "Bob Smith" 
+        category    sth. like economy, sports, ...                                                                                                                                                                                                                           
+        text    the plain text of the article                                                                                                                                                                                                                               
         '''
 
         try:
