+++ rssscrapers/corp_bbva_scraper.py
@@ -20,25 +20,32 @@ def polish(textstring):
     return result.strip()
 
 class bbva(rss):
-    """BBVA"""
+    """scrapes bbva.com"""
 
     def __init__(self,database=True):
         self.database = database
-        self.doctype = "BBVA (corp)"
+        self.doctype = "bbva (corp)"
         self.rss_url ='https://www.bbva.com/en/rss/press-releases'
         self.version = ".1"
         self.date = datetime.datetime(year=2017, month=7, day=5)
 
 
     def parsehtml(self,htmlsource):
-        '''                                                                             
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+        '''                                                                                                                                                                                                                                                                 
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                 
+
+        Parameters                                                                                                                                                                                                                                                       
+        ----                                                                                                                                                                                                                                                               
+        htmlsource: string                                                                                                                                                                                                                                                 
+            html retrived from RSS feed                                                                                                                                                                                                                                     
+
+        yields                                                                                                                                                                                                                                                             
+        ----                                                                                                                                                                                                                                                               
+        title    the title of the article
+        teaser    the intro to the artcile                                                                                                                                                                                                                                   
+        text    the plain text of the article                                                                                                                                                                                                                              
         '''
+
         tree = fromstring(htmlsource)
         try:
             title="".join(tree.xpath('//*/h1[@class="titular"]//text()')).strip()
