+++ rssscrapers/corp_glencore_scraper.py
@@ -31,14 +31,20 @@ class glencore(rss):
 
 
     def parsehtml(self,htmlsource):
-        '''                                                                             
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+        '''                                                                                                                                                                                                                                                                
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                
+        
+        Parameters                                                                                                                                                                                                                                                         
+        ----                                                                                                                                                                                                                                                               
+        htmlsource: string                                                                                                                                                                                                                                                  
+            html retrived from RSS feed                                                                                                                                                                                                                                     
+
+        yields                                                                                                                                                                                                                                                             
+        ----                                                                                                                                                                                                                                                              
+        title    the title of the article                                                                                                                                                                                                                                  
+        text    the plain text of the article                                                                                                                                                                                                                              
         '''
+        
         tree = fromstring(htmlsource)
         try:
             title="".join(tree.xpath('//*[@class="typography"]/h1/text()')).strip()
