+++ rssscrapers/corp_popular_scraper.py
@@ -20,25 +20,32 @@ def polish(textstring):
     return result.strip()
 
 class popular(rss):
-    """Banco Popular"""
+    """Scrapes Banco Popular"""
 
     def __init__(self,database=True):
         self.database = database
-        self.doctype = "Banco Popular (corp)"
+        self.doctype = "banco popular (corp)"
         self.rss_url ='https://www.comunicacionbancopopular.es/feed/?post_type=nota'
         self.version = ".1"
         self.date = datetime.datetime(year=2017, month=8, day=30)
 
 
     def parsehtml(self,htmlsource):
+        '''                                                                                                                                                                                                                                                               
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                 
+
+        Parameters                                                                                                                                                                                                                                                         
+        ----                                                                                                                                                                                                                                                               
+        htmlsource: string                                                                                                                                                                                                                                                 
+            html retrived from RSS feed                                                                                                                                                                                                                                     
+            
+        yields                                                                                                                                                                                                                                                             
+        ----                                                                                                                                                                                                                                                               
+        title    the title of the article                                                                                                                                                                                                                                  
+        teaser    the intro to the artcile                                                                                                                                                                                                                                 
+        text    the plain text of the article                                                                                                                                                                                                                              
         '''
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
-        '''
+
         tree = fromstring(htmlsource)
         try:
             title="".join(tree.xpath('//*/h3[@class="entry-title single-title"]//text()')).strip()
