+++ rssscrapers/corp_unilever_scraper.py
@@ -30,14 +30,22 @@ class unilever (rss):
         self.date = datetime.datetime(year=2017, month=6, day=14)
 
     def parsehtml(self,htmlsource):
+        '''                                                                                                                                                                                                                                                                 
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                 
+
+        Parameters                                                                                                                                                                                                                                                        
+        ----                                                                                                                                                                                                                                                               
+        htmlsource: string                                                                                                                                                                                                                                                 
+            html retrived from RSS feed                                                                                                                                                                                                                                    
+
+        yields                                                                                                                                                                                                                                                             
+        ----                                                                                                                                                                                                                                                                
+        title    the title of the article
+        category    sth. like economy, sports, ...                                                                                                                                                                                                                          
+        teaser    the intro to the artcile                                                                                                                                                                                                                                  
+        text    the plain text of the article                                                                                                                                                                                                                               
         '''
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
-        '''
+
         tree = fromstring(htmlsource)
         try:
             title="".join(tree.xpath('//*/article[@class="content-article"]/h1/text()')).strip()
