+++ rssscrapers/diewelt_scraper.py
@@ -19,7 +19,7 @@ class diewelt(rss):
 
     def __init__(self,database=True):
         self.database=database
-        self.doctype = "Die Welt"
+        self.doctype = "die welt (www)"
         self.rss_url='http://www.welt.de/?service=Rss'
         self.version = ".1"
         self.date    = datetime.datetime(year=2016, month=11, day=21)
