+++ rssscrapers/freitag_scraper.py
@@ -19,7 +19,7 @@ class freitag(rss):
 
     def __init__(self,database=True):
         self.database=database
-        self.doctype = "Der Freitag"
+        self.doctype = "der freitag (www)"
         self.rss_url='https://www.freitag.de/@@RSS'
         self.version = ".1"
         self.date    = datetime.datetime(year=2016, month=11, day=21)
