+++ rssscrapers/gazetvanantwerpen_scraper.py
@@ -29,13 +29,21 @@ class gazetvanantwerpen(rss):
         self.date    = datetime.datetime(year=2017, month=5, day=3)
 
     def parsehtml(self,htmlsource):
-        '''
-        Parses the html source to retrieve info that is not in the RSS-keys
-        In particular, it extracts the following keys (which should be available in most online news:
-        section    sth. like economy, sports, ...
-        text        the plain text of the article
-        byline      the author, e.g. "Bob Smith"
-        byline_source   sth like ANP
+        '''                                                                                                                                                                                                                                                                 
+        Parses the html source to retrieve info that is not in the RSS-keys                                                                                                                                                                                                 
+        
+        Parameters                                                                                                                                                                                                                                                         
+        ----                                                                                                                                                                                                                                                                
+        htmlsource: string                                                                                                                                                                                                                                                 
+            html retrived from RSS feed                                                                                                                                                                                                                                     
+
+        yields                                                                                                                                                                                                                                                              
+        ----                                                                                                                                                                                                                                                                
+        title    the title of the article                                                                                                                                                                                                                                   
+        category    sth. like economy, sports, ...                                                                                                                                                                                                                      
+        text    the plain text of the article                                                                                                                                                                                                                              
+        byline    the author, e.g. "Bob Smith"                                                                                                                                                                                                                             
+        byline_source    sth like ANP                                                                                                                                                                                                                                       
         '''
 
         tree = fromstring(htmlsource)
@@ -78,11 +86,11 @@ class gazetvanantwerpen(rss):
 
         text = textfirstpara + " " + textrest
 
-        extractedinfo={"byline":byline.strip(),
-                       "bylinesource":bylinesource.strip(),
+        extractedinfo={"title":title.strip(),
+                       "category":category.strip(),
                        "text":text.strip(),
-                       "title":title.strip(),
-                       "category":category.strip()
+                       "byline":byline.strip(),
+                       "bylinesource":bylinesource.strip()
                        }
 
         return extractedinfo
