+++ rssscrapers/jungewelt_scraper.py
@@ -18,7 +18,7 @@ class jungewelt(rss):
 
     def __init__(self,database=True):
         self.database = database
-        self.doctype = "https://www.jungewelt.de"
+        self.doctype = "junge welt (www)"
         self.rss_url='https://www.jungewelt.de/feeds/newsticker.rss'
         self.version = ".1"
         self.date    = datetime.datetime(year=2016, month=12, day=28)
