+++ rssscrapers/neuesde_scraper.py
@@ -19,7 +19,7 @@ class neues(rss):
 
     def __init__(self,database=True):
         self.database = database
-        self.doctype = "https://www.neues-deutschland.de"
+        self.doctype = "neues deutschland (www)"
         self.rss_url='https://www.neues-deutschland.de/rss/aktuell.php'
         self.version = ".1"
         self.date    = datetime.datetime(year=2016, month=12, day=28)
